AKAO refactor, can build .mid and .sf2

+++ Assets/VS/FileFormats/AKAO/AKAOInstrument.cs
@@ -7,12 +7,12 @@ namespace VS.FileFormats.AKAO
     public class AKAOInstrument
     {
         public string name = "";
-        public uint program = 0;
+        public byte program = 0;
         public bool isDrum = false;
         public bool a1 = false;
         public AKAORegion[] regions;
 
-        public AKAOInstrument(uint id, bool _isDrum = false)
+        public AKAOInstrument(byte id, bool _isDrum = false)
         {
             program = id;
             isDrum = _isDrum;
